backend/pyproject.toml
[tool.poetry]
name = "pokerstrategy"
version = "0.1.0"
description = ""
authors = ["danka <dchierntsov@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.6"
djangorestframework = "^3.15.1"
markdown = "^3.6"
django-filter = "^24.2"
drf-spectacular = "^0.27.2"
django-cors-headers = "^4.3.1"
psycopg2 = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
